+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added support for Address exports via the native Craft exporter tool.
+- Added [filter fallback](https://www.doublesecretagency.com/plugins/smart-map/docs/using-a-filter-fallback-in-proximity-searches) technique for proximity searches.
 - KML files can now be specified by absolute URL.
 - Added [`field`](https://www.doublesecretagency.com/plugins/smart-map/docs/customizing-the-map-in-twig) option to filter field(s) to be included on map.
 - Added [`hasCoords`](https://www.doublesecretagency.com/plugins/smart-map/docs/filtering-out-entries-with-invalid-coordinates) parameter to filter out results with incomplete coordinates. (thanks @bencroker)